Changed player attribute type from unsigned char to unsigned long.  Breaks savefile compatibility.
Keep track of how much gold the player has collected, spent at shops,
donated, and otherwise used (like at ziggurats) and include that info
in the misc section of the character dump file.

Place-holder for chaos spawn monster.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7683 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/transfor.cc
@@ -386,7 +386,7 @@ bool transform(int pow, transformation_type which_trans, bool quiet)
     }
 
     // This must occur before the untransform() and the is_undead check.
-    if (you.attribute[ATTR_TRANSFORMATION] == which_trans)
+    if (you.attribute[ATTR_TRANSFORMATION] == (unsigned long) which_trans)
     {
         if (you.duration[DUR_TRANSFORMATION] < 100)
         {
